ISSUE-1641: move out table meta ot table_meta.rs
