+++ query/src/catalogs/meta.rs
@@ -13,36 +13,23 @@
 // limitations under the License.
 //
 
-use std::collections::HashMap;
-use std::sync::Arc;
-
 use common_metatypes::MetaId;
 use common_metatypes::MetaVersion;
 
-use crate::catalogs::Table;
-use crate::catalogs::TableFunction;
-
-pub type TableMeta = DatasourceMeta<Arc<dyn Table>>;
-pub type TableFunctionMeta = DatasourceMeta<Arc<dyn TableFunction>>;
-
 #[derive(Debug)]
-pub struct DatasourceMeta<T> {
-    datasource: T,
+pub struct Meta<T> {
+    t: T,
     id: MetaId,
     version: Option<MetaVersion>,
 }
 
-impl<T> DatasourceMeta<T> {
-    pub fn create(datasource: T, id: MetaId) -> Self {
-        Self::with_version(datasource, id, None)
+impl<T> Meta<T> {
+    pub fn create(t: T, id: MetaId) -> Self {
+        Self::with_version(t, id, None)
     }
 
-    pub fn with_version(datasource: T, id: MetaId, version: Option<MetaVersion>) -> Self {
-        Self {
-            datasource,
-            id,
-            version,
-        }
+    pub fn with_version(t: T, id: MetaId, version: Option<MetaVersion>) -> Self {
+        Self { t, id, version }
     }
 
     pub fn meta_ver(&self) -> Option<MetaVersion> {
@@ -54,27 +41,6 @@ impl<T> DatasourceMeta<T> {
     }
 
     pub fn datasource(&self) -> &T {
-        &self.datasource
-    }
-}
-
-pub struct InMemoryMetas {
-    pub(crate) name2meta: HashMap<String, Arc<TableMeta>>,
-    pub(crate) id2meta: HashMap<MetaId, Arc<TableMeta>>,
-}
-
-impl InMemoryMetas {
-    pub fn create() -> Self {
-        InMemoryMetas {
-            name2meta: HashMap::default(),
-            id2meta: HashMap::default(),
-        }
-    }
-
-    pub fn insert(&mut self, tbl_meta: TableMeta) {
-        let met_ref = Arc::new(tbl_meta);
-        self.name2meta
-            .insert(met_ref.datasource().name().to_owned(), met_ref.clone());
-        self.id2meta.insert(met_ref.meta_id(), met_ref);
+        &self.t
     }
 }
