+++ query/src/catalogs/mod.rs
@@ -20,13 +20,17 @@ mod meta;
 mod meta_client;
 mod table;
 mod table_function;
+mod table_meta;
 
 pub use catalog::Catalog;
 pub use database::Database;
 pub use impls::database_catalog::*;
 pub use impls::remote_meta_store_client::RemoteMetaStoreClient;
-pub use meta::*;
+pub use meta::Meta;
 pub use meta_client::MetaClient;
 pub use table::Table;
 pub use table::TablePtr;
 pub use table_function::TableFunction;
+pub use table_meta::InMemoryMetas;
+pub use table_meta::TableFunctionMeta;
+pub use table_meta::TableMeta;
